+++ Release_Notes.md
@@ -27,6 +27,7 @@ Marked AbstractDatabaseFunctions and AbstractBaseModel as Obsolete
     - MFunctionTypeProfile                  to AAddedUpdated
     - MGroupProfile                         to AAddedUpdated
     - MNameValuePair                        to AAddedUpdated
+    - MNameValuePairDetails                 to AAddedUpdated
 
 ### Known Bugs
 - There is an issue where the refresh token is being revoked when it shouldn't be causing a logout to occure